+++ book-network/src/main/java/com/alibou/book/feedback/FeedbackRequest.java
@@ -11,7 +11,7 @@ public record FeedbackRequest(
         @Positive(message = "200")
         @Min(value = 0, message = "201")
         @Max(value = 5, message = "202")
-        Integer note,
+        Double note,
         @NotNull(message = "203")
         @NotEmpty(message = "203")
         @NotBlank(message = "203")
